Added a greedy JSON praser to the JSON view to deal with unparsed JSON in columns
